requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==1.0.0
jwt==0.6.1
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
